+++ log-core/src/commonMain/kotlin/com/ae/log/Lifecycle.kt
@@ -7,31 +7,31 @@ import com.ae.log.core.bus.EventBus
 import com.ae.log.core.bus.PanelClosedEvent
 import com.ae.log.core.bus.PanelOpenedEvent
 
-public class Lifecycle internal constructor(
+internal class Lifecycle(
     private val pluginManager: PluginManager,
     private val eventBus: EventBus,
 ) {
-    public fun notifyStart() {
+    fun notifyStart() {
         pluginManager.forEach { it.onStart() }
         eventBus.publish(AppStartedEvent)
     }
 
-    public fun notifyStop() {
+    fun notifyStop() {
         pluginManager.forEach { it.onStop() }
         eventBus.publish(AppStoppedEvent)
     }
 
-    public fun notifyOpen() {
+    fun notifyOpen() {
         pluginManager.forEach { it.onOpen() }
         eventBus.publish(PanelOpenedEvent)
     }
 
-    public fun notifyClose() {
+    fun notifyClose() {
         pluginManager.forEach { it.onClose() }
         eventBus.publish(PanelClosedEvent)
     }
 
-    public fun clearAll() {
+    fun clearAll() {
         pluginManager.forEach { it.onClear() }
         eventBus.publish(AllDataClearedEvent)
     }